requirements.txt
anyio==3.3.4
arrow==1.1.1
asgiref==3.4.1
binaryornot==0.4.4
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
cmake==3.21.4
cryptography==35.0.0
dlib==19.22.1
dnspython==2.1.0
email-validator==1.1.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.70.0
h11==0.12.0
httptools==0.2.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jinja2-time==0.2.0
MarkupSafe==2.0.1
numpy==1.21.2
orjson==3.6.4
Pillow==8.3.2
poyo==0.5.0
prompt-toolkit==3.0.20
pycparser==2.20
pydantic==1.8.2
PyJWT==2.3.0
pymongo==3.12.1
python-dateutil==2.8.2
python-dotenv==0.19.1
python-multipart==0.0.5
python-slugify==5.0.2
PyYAML==6.0
questionary==1.10.0
requests==2.26.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
text-unidecode==1.3
typer==0.4.0
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.0
